Core-Algorithmn/requirements.txt
Pillow>=9.0.0
flask>=2.0.0
pymongo>=4.0.0
